+++ Changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v0.0.1
+- Added Ray, Camera, Hittable
 
 ## v0.0.0
 (Released: May 20, 2022)